+++ wtb/app/utils/string_process.py
@@ -1,7 +1,7 @@
 import re
 
 import bs4
-from data_structures.task_board import FunctionAssignment, TaskBoard
+from data_structures.taskboard import FunctionAssignment, TaskBoard
 
 
 def empty_cell(cell_value: str) -> bool:
